move patches to appropriate places
