+++ SolastaCommunityExpansion/Models/CeContentPackContext.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Reflection;
 using HarmonyLib;
 using SolastaCommunityExpansion.Builders;
 using SolastaCommunityExpansion.Utils;
@@ -11,6 +10,22 @@ namespace SolastaCommunityExpansion.Models
 {
     internal static class CeContentPackContext
     {
+        // Works in conjuction with "custom resources enablement patch" (patches are marked with this comment)
+        internal class CEAssetReferenceSprite : AssetReferenceSprite
+        {
+            public CEAssetReferenceSprite(Sprite sprite) : base(string.Empty)
+            {
+                Sprite = sprite;
+            }
+
+            public Sprite Sprite { get; }
+            public override UnityEngine.Object Asset => Sprite;
+            public override bool RuntimeKeyIsValid()
+            {
+                return true;
+            }
+        }
+
         internal const GamingPlatformDefinitions.ContentPack CeContentPack = (GamingPlatformDefinitions.ContentPack)9999;
 
         internal static ContentPackDefinition ContentPackDefinition = CreateContentPackDefinition();
@@ -42,68 +57,6 @@ namespace SolastaCommunityExpansion.Models
             {
             }
             #endregion 
-        }
-    }
-
-
-    // TODO: move to appropriate place(s)
-
-    // Works in conjuction with 2 patches below
-    class CEAssetReferenceSprite : AssetReferenceSprite
-    {
-        public CEAssetReferenceSprite(Sprite sprite) : base(string.Empty)
-        {
-            Sprite = sprite;
-        }
-
-        public Sprite Sprite { get; }
-        public override UnityEngine.Object Asset => Sprite;
-        public override bool RuntimeKeyIsValid()
-        {
-            return true;
-        }
-    }
-
-    [HarmonyPatch]
-    internal static class Gui_LoadAssetAsync
-    {
-        internal static MethodBase TargetMethod()
-        {
-            return AccessTools.Method(typeof(Gui), "LoadAssetSync", new Type[] { typeof(AssetReference) }, new Type[] { typeof(Sprite) });
-        }
-
-        internal static bool Prefix(AssetReference asset)
-        {
-            // If it's a CEAssetReferenceSprite prevent async load
-            return asset is not CEAssetReferenceSprite;
-        }
-
-        internal static void Postfix(AssetReference asset, ref Sprite __result)
-        {
-            if(asset is CEAssetReferenceSprite reference)
-            {
-                Main.Log($"Providing sprite {reference.Sprite.name}");
-
-                // Return our sprite
-                __result = reference.Sprite;
-            }
-        }
-    }
-
-    [HarmonyPatch(typeof(Gui), "ReleaseAddressableAsset")]
-    internal static class Gui_ReleaseAddressableAsset
-    {
-        internal static bool Prefix(UnityEngine.Object asset)
-        {
-            // If it's a CE provided sprite stop it being unloaded
-            bool retval = !CustomIcons.IsCachedSprite(asset as Sprite);
-
-            if (!retval)
-            {
-                Main.Log($"Not releasing {asset.name}");
-            }
-
-            return retval;
         }
     }
 }
